refactor(pm-quality): isolate command selection model (#596)
